Change all float to double

+++ ggEngine/ggEngine/Add.cpp
@@ -11,7 +11,7 @@ namespace ggEngine{
 		this->device = drawManager->GetDevice();
 		this->world = world;
 	}
-	Sprite* Add::Sprite(float x, float y, std::string textureKey, ggEngine::Group *group){
+	Sprite* Add::Sprite(double x, double y, std::string textureKey, ggEngine::Group *group){
 		SpriteInfo* inf = this->cache->GetSpriteInfo(textureKey);
 		ggEngine::Sprite *sprite = new ggEngine::Sprite(this->drawManager, inf);
 		sprite->SetParentObject(group);
@@ -19,7 +19,7 @@ namespace ggEngine{
 		group->AddDrawObjectToList(sprite);
 		return sprite;
 	}
-	SpriteAnimation* Add::SpriteAnimation(float x, float y, std::string textureKey, int frameWidth, int frameHeight, ggEngine::Group * group, int defaultFrame, int numberOfFrame)
+	SpriteAnimation* Add::SpriteAnimation(double x, double y, std::string textureKey, int frameWidth, int frameHeight, ggEngine::Group * group, int defaultFrame, int numberOfFrame)
 	{
 		SpriteInfo* inf = this->cache->GetSpriteInfo(textureKey);
 		ggEngine::SpriteAnimation *spriteAnimation = new ggEngine::SpriteAnimation(this->drawManager, inf, frameWidth, frameHeight, defaultFrame, numberOfFrame);
@@ -33,7 +33,7 @@ namespace ggEngine{
 		world->AddGroup(gr);
 		return gr;
 	}
-	Text* Add::Text(float x, float y, std::string fontKey, float width, float height, std::string text, Style style, ggEngine::Group  *group)
+	Text* Add::Text(double x, double y, std::string fontKey, double width, double height, std::string text, Style style, ggEngine::Group  *group)
 	{
 		Font* font = this->cache->GetFont(fontKey);
 		ggEngine::Text *textObject = new ggEngine::Text(this->drawManager,font, x, y, width, height, text, style);
