+++ ggEngine/ggEngine/Box.h
@@ -2,9 +2,9 @@
 #include <d3d9.h>
 namespace ggEngine {
 	struct Box {
-		float x, y;
-		float w, h;
-		float vx, vy;
+		double x, y;
+		double w, h;
+		double vx, vy;
 		RECT GetRect() { return RECT{x,y,x+w,y+h}; }
 	};
 }
