+++ ggEngine/ggEngine/Circle.cpp
@@ -1,6 +1,6 @@
 #include "Circle.h"
 
-ggEngine::Circle::Circle(float radius)
+ggEngine::Circle::Circle(double radius)
 {
 	this->radius = radius;
 	this->pCenter = Vector(radius, radius);
@@ -10,7 +10,7 @@ ggEngine::Circle::~Circle()
 {
 }
 
-float ggEngine::Circle::GetArea()
+double ggEngine::Circle::GetArea()
 {
 	return MATH_PI*radius*radius;
 }
