+++ ggEngine/ggEngine/ConstantEnum.h
@@ -7,8 +7,8 @@
 #define VECTOR_ZERO (Vector(0,0))
 #define DEFAULT_IMAGE ""
 #define D3DFVF_TL (D3DFVF_XYZRHW | D3DFVF_DIFFUSE | D3DFVF_TEX1)
-#define GRAVITY (float(9.81))
-#define MATH_PI (float(3.41))
+#define GRAVITY (double(9.81))
+#define MATH_PI (double(3.41))
 #define DEFAULT_COLOR D3DCOLOR_RGBA(170, 255, 0, 150)
 #define PIXEL_PER_CENTIMETER 100
 const bool PRINT_DETAIL = false;
